Switched to int64, moved logsumexp to utils.
